requirements.txt
python-dotenv==1.0.1
homeharvest==0.3.28
langchain==0.2.1
langchain-openai==0.1.8
langchainhub==0.1.17
chainlit==1.1.202